Fix bug in compress.py causing an incorrect compressed output.
I think it was trying to be clever and allow a back-reference to reference
itself, but that only works if the decoder goes byte by byte. Mine does not.

Interestingly, while it seems that bit of "cleverness" would make a smaller
output, it actually did not. The re-compressed veresion is 43 bytes smaller.